common-v2.3.2/crypto/pkcs11/rand.go
/*
Copyright (C) BABEC. All rights reserved.
Copyright (C) THL A29 Limited, a Tencent company. All rights reserved.

SPDX-License-Identifier: Apache-2.0
*/

package pkcs11

import (
	"errors"
	"fmt"
)

const (
	defaultRandomLen = 6
)

func GenerateOTP(p11 *P11Handle, length int) (string, error) {
	if length <= 0 {
		length = defaultRandomLen
	}
	randBytes, err := GenerateBytesOTP(p11, length)
	if err != nil {
		return "", err
	}
	r, err := bytesToInt(randBytes)
	if err != nil {
		return "", err
	}
	rStr := fmt.Sprintf("%0*d", length, r)
	if len(rStr) < length {
		return "", errors.New("generate random failed, len is too low")
	}
	return rStr[:length], nil
}

func GenerateBytesOTP(p11 *P11Handle, length int) ([]byte, error) {
	if length <= 0 {
		length = defaultRandomLen
	}
	return p11.GenerateRandom(length)
}
